TestContest: some doubles replaced with unsigned type

+++ Common/Utils/TestContest/src/ReplayLogger.h
@@ -15,7 +15,7 @@
 
 class CReplayLogger {
 public:
-  typedef void (*FGPSHandler)(void *user, double time, double latitude, double longitude, double altitude);
+  typedef void (*FGPSHandler)(void *user, unsigned time, double latitude, double longitude, unsigned altitude);
   
 private:
   class CCatmullRomInterpolator;
